refactor(app): isolate analysis cache policy
